When a voter signs in with Twitter or Facebook, we now take the voter back to the page where the sign in was initiated. Sign in with email not updated yet. Added profile photo and drop-down to Welcome page.
