+++ src/js/components/Navigation/HeaderBackToBallot.jsx
@@ -330,14 +330,15 @@ class HeaderBackToBallot extends Component {
 
           {this.state.profilePopUpOpen && voter.is_signed_in && (
           <HeaderBarProfilePopUp
-            {...this.props}
+            hideProfilePopUp={this.hideProfilePopUp}
             onClick={this.toggleProfilePopUp}
             profilePopUpOpen={this.state.profilePopUpOpen}
-            weVoteBrandingOff={this.state.we_vote_branding_off}
-            toggleProfilePopUp={this.toggleProfilePopUp}
-            hideProfilePopUp={this.hideProfilePopUp}
-            transitionToYourVoterGuide={this.transitionToYourVoterGuide}
             signOutAndHideProfilePopUp={this.signOutAndHideProfilePopUp}
+            toggleProfilePopUp={this.toggleProfilePopUp}
+            toggleSignInModal={this.toggleSignInModal}
+            transitionToYourVoterGuide={this.transitionToYourVoterGuide}
+            voter={this.props.voter}
+            weVoteBrandingOff={this.state.we_vote_branding_off}
           />
           )}
