+++ src/js/routes/Process/FacebookSignInProcess.jsx
@@ -1,10 +1,12 @@
 import { Component } from 'react';
+import AppActions from '../../actions/AppActions';
 import FacebookActions from '../../actions/FacebookActions';
 import FacebookStore from '../../stores/FacebookStore';
 import { historyPush } from '../../utils/cordovaUtils';
 import LoadingWheel from '../../components/LoadingWheel';
 import { renderLog } from '../../utils/logging';
 import VoterActions from '../../actions/VoterActions';
+import cookies from '../../utils/cookies';
 // This will be needed in the future
 // import WouldYouLikeToMergeAccounts from "../../components/WouldYouLikeToMergeAccounts";
 
@@ -63,17 +65,31 @@ export default class FacebookSignInProcess extends Component {
       // Prevent voterMergeTwoAccountsByFacebookKey from being called multiple times
       this.setState({ mergingTwoAccounts: true });
     }
+    let redirectPathname;
+    const signInStartPath = cookies.getItem('sign_in_start_path');
     if (voterHasDataToPreserve) {
+      redirectPathname = '/more/network';
+      if (signInStartPath) {
+        redirectPathname = signInStartPath;
+        AppActions.unsetStoreSignInStartPath();
+        cookies.removeItem('sign_in_start_path', '/');
+      }
       historyPush({
-        pathname: '/more/network',
+        pathname: redirectPathname,
         state: {
           message: 'Your accounts have been merged.',
           message_type: 'success',
         },
       });
     } else {
+      redirectPathname = '/ballot';
+      if (signInStartPath) {
+        redirectPathname = signInStartPath;
+        AppActions.unsetStoreSignInStartPath();
+        cookies.removeItem('sign_in_start_path', '/');
+      }
       historyPush({
-        pathname: '/ballot',
+        pathname: redirectPathname,
         query: { wait_until_voter_sign_in_completes: 1 },
         state: {
           message: 'You have successfully signed in with Facebook.',
@@ -86,8 +102,15 @@ export default class FacebookSignInProcess extends Component {
   voterFacebookSaveToCurrentAccount () {
     // console.log("In voterFacebookSaveToCurrentAccount");
     VoterActions.voterFacebookSaveToCurrentAccount();
+    let redirectPathname = '/friends';
+    const signInStartPath = cookies.getItem('sign_in_start_path');
+    if (signInStartPath) {
+      redirectPathname = signInStartPath;
+      AppActions.unsetStoreSignInStartPath();
+      cookies.removeItem('sign_in_start_path', '/');
+    }
     historyPush({
-      pathname: '/friends',
+      pathname: redirectPathname,
       state: {
         message: 'You have successfully signed in with Facebook.',
         message_type: 'success',
