+++ src/sass/components/_device-menu.scss
@@ -3,6 +3,7 @@
 
 
 .header-slide-out-menu-text-left {
+  color: $black;
   font-size: 1.2rem;
 }
